logic was separated to repository and service classes
